[Vulkan][Tests] Merged test_target_codegen_spirv.py into test_target_codegen_vulkan.py
